requirements.txt
tornado
sphinx
Pillow
nose
logilab-astng
coverage
python-xmp-toolkit
